# Rare_Disease_Atlas_Backend_V2.2/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rare Disease Atlas public read API",
    "version": "1.0.0",
    "description": "Seven versioned product routes; two operational routes. Public, read-only, no browser credentials. Synthetic fixtures are not medical evidence."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development; configure the real HTTPS host after deployment."
    }
  ],
  "security": [],
  "paths": {
    "/v1/meta": {
      "get": {
        "operationId": "getMeta",
        "summary": "Read meta from the pinned snapshot",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "get": {
        "operationId": "getSearch",
        "summary": "Read search from the pinned snapshot",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contexts/{id}": {
      "get": {
        "operationId": "getContext",
        "summary": "Read context from the pinned snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 180
            },
            "description": "Opaque identifier. URL-encode once; do not split on colons."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contexts/{id}/connections": {
      "get": {
        "operationId": "getConnections",
        "summary": "Read connections from the pinned snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 180
            },
            "description": "Opaque identifier. URL-encode once; do not split on colons."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assertions/{id}": {
      "get": {
        "operationId": "getAssertion",
        "summary": "Read assertion from the pinned snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 180
            },
            "description": "Opaque identifier. URL-encode once; do not split on colons."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssertionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calculations/{id}": {
      "get": {
        "operationId": "getCalculation",
        "summary": "Read calculation from the pinned snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 180
            },
            "description": "Opaque identifier. URL-encode once; do not split on colons."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contexts/{id}/actions": {
      "get": {
        "operationId": "getActions",
        "summary": "Read actions from the pinned snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 180
            },
            "description": "Opaque identifier. URL-encode once; do not split on colons."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Snapshot unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Warning": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "request_id",
              "retryable"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "error"
        ],
        "additionalProperties": false
      },
...
      "MetaResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "type": "string"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "real",
              "synthetic_fixture"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/MetaData"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            }
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "data_mode",
          "data",
          "warnings"
        ],
        "additionalProperties": false
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "type": "string"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "real",
              "synthetic_fixture"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SearchData"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            }
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "data_mode",
          "data",
          "warnings"
        ],
        "additionalProperties": false
      },
      "ContextResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "type": "string"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "real",
              "synthetic_fixture"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ContextData"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            }
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "data_mode",
          "data",
          "warnings"
        ],
        "additionalProperties": false
      },
      "ConnectionsResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "type": "string"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "real",
              "synthetic_fixture"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ConnectionsData"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            }
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "data_mode",
          "data",
          "warnings"
        ],
        "additionalProperties": false
      },
      "AssertionResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "type": "string"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "real",
              "synthetic_fixture"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/AssertionData"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            }
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "data_mode",
          "data",
          "warnings"
        ],
        "additionalProperties": false
      },
      "CalculationResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "const": "1.0.0"
          },
          "snapshot_id": {
            "type": "string"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "real",
              "synthetic_fixture"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/CalculationData"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            }
          }
        },
        "required": [
          "contract_version",
          "snapshot_id",
          "data_mode",
          "data",
          "warnings"
        ],
        "additionalProperties": false
      },
...
    }
  }
}